requirements.txt
playwright
python-chess
configparser